my first commit

+++ 0x00-hello_world/6-size.c
@@ -16,10 +16,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: %zu byte(s)\n", sizeof(a));
+printf("size of a int: %zu byte(s)\n", sizeof(b));
+printf("size of a long int: %zu byte(s)\n", sizeof(c));
+printf("size of a long long int: %zu byte(s)\n", sizeof(d));
+printf("size of a float: %zu byte(s)\n", sizeof(f));
 return (0);
 }
